Add files via upload
This file is an improved version of the first 'Exomoon Code.ipybn' file.

The improvement is calculating each star's luminosity using L =4πR²σT⁴ → the surface luminosity of a given body with radius, R, and effective temperature, T and σ being the Stefan-Boltzmann constant.

This small calculation actually increased the amount of filtered exoplanets producing 18 planets rather than 15. This could be due to missing stellar luminosity values in the .csv file, with those having been initially removed during the filtering process.