Add files via upload
This piece of code builds on the previous file, 'Exomoon Code MK2.ipybn' where I test on Earth-sized planets 0.5 <= R🜨 <= 2.0 that either has been discovered by the transit method/transit timing variation method (TTV), flagged as a transit/TTV, or has transiting data such as transit depth, duration or midpoint. All of these Earth-sized planets are within the habitable zone of their respective stars. 

The reason for this step is to compare my filtered list to a given list of Earth-sized planets that are within the habitable zone and to see if there are similiar planets (i.e. Trappist-1 planets). The list that the code is compared to is from table 1 from the Planetary Habitability Laboratory @ UPR Arecibo (https://phl.upr.edu/hwc/data).