This file is a more concise notebook where it tests for detectability.
Using the cells in the previous file 'Light_Curve_Simulations_160924.pynb', I simulate the light curves using the parameters for 'N' amount of observations and calculate the noise for each light curve.

The new light curves based off of the randomised angles in the orbit simulation and TTV calculations to plot a histogram of fitted 'time of inferior conjunction' times  (t_0) and their uncertainties. This is then used to calculate the probability of observing a TTV from 'N' observations.